Refactor start and save popups into separate modules
